LayoutTests/svg/animations/animate-calcMode-spline-by-expected.txt

CONSOLE MESSAGE: line 2: ReferenceError: Can't find variable: createSVGTestCase
SVG 1.1 dynamic animation tests

Test calcMode spline with by animation. You should see a green 100x100 rect and only PASS messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
